Consolidate utility functions
Some of these are defined in multiple places so it makes sense to put
them in a single module.
