+++ src/nexpy/examples/plugins/chopper/get_ei.py
@@ -1,6 +1,6 @@
 import numpy as np
 from nexpy.gui.datadialogs import BaseDialog, GridParameters
-from nexpy.gui.mainwindow import report_error
+from nexpy.gui.utils import report_error
 from nexusformat.nexus import NeXusError
